fix: 完成Discord机器人相关修复和优化
